+++ peregrine-core/src/shape/core.rs
@@ -46,8 +46,8 @@ pub(super) fn bulk<T>(b: Vec<T>, a_len: usize, primary: bool) -> Vec<T> where T:
 #[derive(Clone,Debug,PartialEq,Eq,Hash)]
 pub struct DirectColour(pub u8,pub u8,pub u8);
 
-#[derive(Clone,Debug)]
-pub struct Pen(pub String,pub f64,pub Vec<DirectColour>);
+#[derive(Clone,Debug,PartialEq,Eq,Hash)]
+pub struct Pen(pub String,pub u32,pub Vec<DirectColour>);
 
 impl Pen {
     pub fn bulk(self, len: usize, primary: bool) -> Pen {
